worker/pyproject.toml
[tool.poetry]
name = "worker"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = {extras = ["email"], version = "^2.7.1"}
pydantic-settings = "^2.2.1"
motor = "^3.4.0"
aiormq = "^6.8.0"
jinja2 = "^3.1.3"
faker = "^25.0.0"
aiohttp = "^3.9.5"
backoff = "^2.2.1"
python-http-client = "^3.3.7"
sendgrid = "^6.11.0"
websocket-client = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
